Trunk: Reintegration merged EXP-namelist at r29745 to the trunk. This update brings changes from two sources, EXP-namelist and mpondeca6.
From EXP-namelist -

A new option has been added to tlnmc_option:

 * tlnmc_option=4 : if in EnVar mode, apply constraint to static contribution ONLY

Now, tlnmc_option=1 is ONLY valid for 3DVAR mode. Additionally, added logic so that if a user selects a tlnmc_option between 2 and 4, but are in 3DVAR mode, assume they intend to run with the constraint (so reset it to 1).

The EnKF innovation scripts have been added for both ecf and sms to the scripts directory.

Finally, removed the slow version of the tlnmc code from the GSI and make modifications to the tlnmc_type namelist option (removing from global and replacing with reg_tlnmc_type for regional).

From mpondeca6 -

For the option "twodvar_regional=.true." only:

 * Reverse a recent change to gridmod.F90 as it leads to incorrect results when FGAT is turned on.
 * Update function dplev_mask in projmethod_support.f90, so that the optional biorthogonalization of the gradient vectors can still work as new analysis variables are added to the system.



git-svn-id: https://svnemc.ncep.noaa.gov/projects/gsi/trunk/scripts@29752 281fb692-3a18-0410-826d-9e847983f365