Trunk: Merged wwu_vvlocal branch at r42609 to the trunk. The ability is available for the global but not for the regional applications. The background error covariances for alpha is applied in spectral space for the global but is with recursive filters for regional. An extra dimension will be added to the localization related variables. The normalization of the recursive filter for each vertical level has to be done separately. As in the global application, the localization scales will be read in from an input file. Most of the changes will be done in the subroutine hybrid_ensemble_isotropic.F90. The memory requirement will be the same if the localization scales are constant in the vertical, but will increase 18% if the scales are vertically varying.
The option of constant localization in the vertical which has lower memory requirement than the vertically varying ones has been kept. To turn on the vertical varying option, set the input parameter in namelist HYBRID_ENSEMBLE

    readin_localization=.true.

and provide the file 

    hybens_locinfo

The results before and after the changes if the vertically varying option was turn off are reproducible. The same results were also obtained when the option was on and the scales specified in hybens_locinfo are the same as the single scale specified in the input parameter.

Further, corrected type variables from grd_ens to grd_loc in a few places.

Additional bug fixes with respect to updating cloud variables were made to wrwrfnmma.F90 and read_wrf_nmm_guess.F90.

Deallocated arrays in read_seviri.f90.

* fix:
  * nam_hybens_d01_locinfo - added

* src:
  * read_wrf_nmm_guess.F90 - detach use_reflectivity from n_actual_clouds
  * update_guess.f90 - remove unused variable, use_reflectivity
  * read_gfs_ozone_for_regional.f90 - comment added for bug fix for undefined variable "proceed" in check_vars_
  * hybrid_ensemble_parameters.f90 - add logical variable vvlocal for vertically verying horizontal localization length in regional
  * hybrid_ensemble_isotropic.F90 - modifications to allow vertically varying scales
  * read_seviri.f90 - deallocated three arrays (datasev1, datasev2, and hdr)
  * wrwrfnmma.F90 - detach use_reflectivity from n_actual_clouds
  * gsi_nemsio_mod.f90 - remove debugging printout


git-svn-id: https://svnemc.ncep.noaa.gov/projects/gsi/trunk/scripts@42952 281fb692-3a18-0410-826d-9e847983f365