Trunk: Merged EXP-enkfnemsiofixes branch at r47941 to the trunk. An option to read and write NEMSIO files for GFS and NMMB was recently added to the EnKF code.  This update fixes errors in the conversion of sensible temperature to virtual temperature, and vice versa, in the NEMSIO I/O routines (discovered by Craig Schwartz at NCAR).  The same bug was fixed for WRF ARW and NMM, but it turns out that the virtual temperature is only used for WRF in the computation of the saturation specific humidity used to define a normalized humidity analysis variable when pseudo_rh = .true.
The Makefiles were also modified so that EXE_FILE is set in Makefile.conf, not Makefile.  This makes is easier to create separate executables for GFS, WRF and NMMB without changing Makefile.

Finally, an update was made to general_write_gfsatm.f90 to ensure that the 4D ensemble increments are reproducible as the number of processors change.



git-svn-id: https://svnemc.ncep.noaa.gov/projects/gsi/trunk/scripts@47948 281fb692-3a18-0410-826d-9e847983f365