Trunk: Merged EXP-efmn branch at r76646 to the trunk.
A failure of an ensemble member in operations (or global parallels) implies that all members following the failed member within that group will not be integrated. This can give a false sense of how many members actually failed. A work-around is proposed to allow the members to complete integration, whilst keeping a tab on which members actually failed.
 * If a few (say 2-3) fail, one can regenerate those through a linear combination of successful members.
 * If many (say 10) fail, one can regenerate by adding lagged perturbations akin to a cold start of the failed members.

Please see ticket #498 for additional details.

Updates:
 * scripts/EnKF/scripts_ncep/exglobal_enkf_fcst.sh.ecf - updated to allow all members to integrate even if one fails. Trigger *hard* failure at the end.


git-svn-id: https://svnemc.ncep.noaa.gov/projects/gsi/trunk/scripts@76662 281fb692-3a18-0410-826d-9e847983f365